[scripts] toggle compton

+++ bin/scripts/toggle_comp.sh
@@ -1,3 +1,6 @@
 #!/bin/sh
-killall compton 
-compton -b --config ${XDG_CONFIG_HOME}/compton/compton.conf &
+if [[ $(pidof compton) ]]; then
+    killall compton 
+else
+    compton -b --config ${XDG_CONFIG_HOME}/compton/compton.conf &
+fi
